requirements.txt
openai
aiohttp
gradio
python-dotenv
